updated readme and cleaned up examples
