+++ README.md
@@ -395,7 +395,7 @@ Note: As of 1.1.3 Generators now use IIncrementalGenerator interface and only ge
 ```xml
 <ProjectReference Include="..\..\source\WebLogger.Generators\WebLogger.Generators.csproj" />
 ```
-After adding the package the `CommandHandlerAttribute` will be generated and added to your project in the CommandHandlerAttribute.g.cs file.
+After adding the package the `CommandHandlerAttribute`, `CommandStore`, `TargetCommand`, `IStoredCommands`, and `StoredCommandsExtensions` types will be generated and added to your project in the CommandHandlerAttribute.g.cs file.
 
 ```csharp
 // <auto-generated/>
@@ -472,6 +472,122 @@ logger.DiscoverCommands(Assembly.GetAssembly(typeof(Program)))
 // Or create a new instance and register.
 
 logger.RegisterCommand(new GeneratedCommand());
+```
+You can also create a single class containing several CLI commands known as a `CommandStore`.  Create a partial class
+marked with the `CommandStore` attribute that includes at least one method marked with the `TargetCommand` attribute.  The source generator will created a new partial class extending your functionality.
+This new partial class will include a private readonly `List<IWebLoggerCommand> _commands` pointing the handlers at the methods you defined.
+
+```csharp
+[CommandStore]
+public partial class RoomControlCommandStore
+{
+    [TargetCommand(
+        "PWRON",
+        "Powers on the room",
+        "Parameters: NA")]
+    public ICommandResponse PowerOnCommand(string command, List<string> args)
+    {
+        //Power on the room and return the results.
+        return CommandResponse.Success(command, "We did the thing worth doing");
+    }
+
+    [TargetCommand(
+        "PWROFF",
+        "Powers off the room",
+        "Parameters: NA")]
+    public ICommandResponse PowerOffCommand(string command, List<string> args)
+    {
+        //Power off the room and return the results.
+        throw new NotImplementedException("Room power off threw exception");
+    }
+    
+    [TargetCommand(
+        "SOURCE",
+        "Selects the specified source in the room",
+        "Parameters: NA")]
+    public ICommandResponse SelectSourceCommand(string command, List<string> args)
+    {
+        return args.Count > 0 
+            ? CommandResponse.Success(command, $"Selected source: {args[0]}") 
+            : CommandResponse.Failure(command, "Please provide a valid Source key argument.");
+    }
+
+    /// <summary>
+    /// Note, this method will not generate a command.
+    /// </summary>
+    public void TestInvalid()
+    {
+
+    }
+}
+
+```
+The generator will implement the `IStoredCommands` interface and you type can now be used to register commands with the weblogger.  
+During command registration your private list will be constructed and populated with values.
+```csharp
+Log.Logger = new LoggerConfiguration()
+    .MinimumLevel.Verbose()
+    .WriteTo.WebloggerSink(
+        logger =>
+        {
+            //Use the generated logger extension method
+            logger.RegisterCommandStore(new RoomControlCommandStore());
+
+            //Or use the IStoredCommands interface extensions
+            var roomCli = new RoomControlCommandStore();
+            roomCli.RegisterCommands(logger);
+        })
+    .WriteTo.Console()
+    .CreateLogger();
+```
+Behind the scenes the follow partial class will created by the generator.
+```csharp
+// <auto-generated/> @6/6/2023 7:59:59 PM
+namespace WebLogger.ConsoleApp.GeneratedCommandStore
+{
+    [global::System.CodeDom.Compiler.GeneratedCodeAttribute("WebLogger", "1.1.4")]
+    partial class RoomControlCommandStore : global::WebLogger.Generators.IStoredCommands
+    {
+        //Code generated by reading the following:
+        //Namespace : namespace WebLogger.ConsoleApp.GeneratedCommandStore
+        //Name: RoomControlCommandStore
+
+        private List<IWebLoggerCommand> _commands;
+
+        private ICommandResponse PowerOnCommand_Generated(string command, List<string> args)
+        {
+            try
+            {
+                return PowerOnCommand(command, args);
+            }
+            catch (Exception e)
+            {
+                return CommandResponse.Error(command, e.Message);
+            }
+        }
+        private ICommandResponse PowerOffCommand_Generated(string command, List<string> args)
+        {
+            try
+            {
+                return PowerOffCommand(command, args);
+            }
+            catch (Exception e)
+            {
+                return CommandResponse.Error(command, e.Message);
+            }
+        }
+
+        public IEnumerable<IWebLoggerCommand> GetStoredCommands()
+        {
+            _commands = new List<IWebLoggerCommand>()
+            {
+                new WebLoggerCommand(PowerOnCommand_Generated, "PWRON","Powers on the room","Parameters: NA"),
+                new WebLoggerCommand(PowerOffCommand_Generated, "PWROFF","Powers off the room","Parameters: NA"),
+            };
+            return _commands;
+        }
+    }
+}
 ```
 
 There are still a *few open issues* that could really improve this.
