+++ source/WebLogger.Generators/IncrementalCommandStoreGenerator.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Net.Mime;
 using System.Runtime.CompilerServices;
 using System.Text;
 using System.Threading;
